atari-ddqn/requirements.txt
numpy
gym
keras
matplotlib
opencv-python
tensorflow-gpu
Pillow
gym[atari]
